move set_pytorch_cuda_alloc_conf to a different module to have fewer loaded dependencies for the CLI
